+++ tests/haskell/Numeric/LinearAlgebra/Coerce/Main.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+
+import Compiler.Prim (coerce)
+import Numeric.LinearAlgebra
+
+bad :: Matrix Int -> Matrix Double
+bad = coerce